실시간 랭킹 업데이트 시간 한국 시간으로 변경 및 monthlyPopularityScore 타입 int에서 double로 변경

See merge request s10-bigdata-recom-sub2/S10P22A306!223

+++ back/src/main/java/site/chachacha/fitme/domain/product/entity/Product.java
@@ -73,7 +73,7 @@ public class Product extends BaseEntity {
 
     private int likeCount = 0;
 
-    private int monthlyPopularityScore = 0;
+    private double monthlyPopularityScore = 0.0;
 
 //    private int weeklyPopularityScore = 0;
